Move letterbox function to utils.py
